get_box_color
removed possibility for non-integer values to be returned

+++ TwitterSearch.py
@@ -26,9 +26,9 @@ class Politician:
 
     def get_box_color(self):
         if self.aff_scale < 0:
-            return (255*(1+self.aff_scale), 255*(1+self.aff_scale), 255)
+            return (int(255*(1+self.aff_scale)), int(255*(1+self.aff_scale)), 255)
         elif self.aff_scale > 0:
-            return (255, 255*(1-self.aff_scale), 255*(1-self.aff_scale))
+            return (255, int(255*(1-self.aff_scale)), int(255*(1-self.aff_scale)))
         else:
             return (255, 255, 255)
